Added version with different task passing method
In previous task passing agents could get stuck simply because two
agents could not swap tasks, both being busy. In the
Functions/passTasks_v2.m a new procedure is defined in which this does
not happen. Problem here is that occasionally tasks are returned to the
common pool and randomly re-assigned to agents during processing, which
is not very nice.

In this commit a bug in passTasks_v2.m was corrected because in finding
similar agents a < was used for comparison instead of <= so when the
threshold was set to 0 agents were not similar even to themselves.

Also, a bug in Functions/GenTask.m was eliminated, namely, I forgot to
normalize after removal of subtasks that no agent could solve.