Adding communication cost
Implemented communication cost: when an agent receives a task it gets
into a refractroy state whan it cannot start working on the task for a
given amount of time. It can still pass the task if need comes. This is
implemented in Functions/PassingSolveTasks.m and
Scripts/ATRAN_SimpleComm.m was modified to include the cost parameter.